Add Kubernetes Deployment and Service for JetBrains Guide
This commit introduces Kubernetes Deployment and Service for the JetBrains Guide application. The deployment.yaml file indicates that two replicas of the application will run in the `jetbrains-guide` namespace. The NGINX configuration is extracted into a ConfigMap to allow more flexible tuning and makes use of the `nginx-conf-cm` ConfigMap. The service.yaml file creates a new service that exposes the application on port 30004.

Dockerfile has been simplified by removing unnecessary steps and introducing a multi-stage build process. This change aims to leverage Kubernetes for configuration management and streamline the build process.