Django speed up query tip
This is a guide article to complement this YouTube short: https://youtube.com/shorts/mewTap8ZQ6w?feature=share
The short is currently "unlisted" but as we have a link, this is viewable in the guide.
We can publish the guide whenever we want, I will make the YouTube short public on Wednesday.

Merge-request: JETBRAINS-GUIDE-MR-391
Merged-by: Sarah Boyce <sarah.boyce@jetbrains.com>